requirements.txt
altair==4.2.0
attrs==22.2.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cryptography==39.0.0
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
docx2txt==0.8
entrypoints==0.4
filelock==3.9.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
mdurl==0.1.2
numpy==1.24.1
packaging==23.0
pandas==1.5.3
pdfminer.six==20221105
pdfplumber==0.7.6
Pillow==9.4.0
pipenv==2022.12.19
platformdirs==2.6.2
plotly==5.13.0
protobuf==3.20.3
pyarrow==10.0.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
PyMySQL==1.0.2
pyparsing==3.0.9
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rich==13.2.0
scipy==1.10.0
seaborn==0.12.2
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.17.0
streamlit-aggrid==0.3.3
tenacity==8.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
Wand==0.6.11
watchdog==2.2.1
zipp==3.11.0
